requirements.txt
# flask==2.3.3
# flask_sqlalchemy==3.1.1
# flask_login==0.6.3
# flask_bcrypt==1.0.1
# wtforms==3.1.2
# transformers==4.35.2
# torch>=2.0.0
# spacy==3.7.2
# nltk==3.8.1
# requests==2.31.0
# python-dotenv


flask==2.3.3
flask_sqlalchemy==3.1.1
flask_login==0.6.3
flask_bcrypt==1.0.1
flask-wtf==1.1.1
flask-migrate==4.0.5
wtforms==3.1.2
transformers==4.35.2
torch>=2.0.0
spacy==3.7.2
nltk==3.8.1
pandas==2.1.4
requests==2.31.0
python-dotenv
textblob==0.17.1